+++ src/game/GameObject.hpp
@@ -7,15 +7,15 @@
 class GameObject {
 	protected:
 	std::shared_ptr<GameObjectRenderer> m_renderer = 0;
-	int m_iX = 0;
-	int m_iY = 0;
+	float m_iX = 0.0f;
+	float m_iY = 0.0f;
 
 	public:
 	virtual ~GameObject() {}
 	virtual void init() {};
-	virtual void setPos(int x, int y);
-	virtual int getX() const;
-	virtual int getY() const;
+	virtual void setPos(float x, float y);
+	virtual float getX() const;
+	virtual float getY() const;
 	virtual void render();
 };
